Added password verification to the launcher

+++ utils/password.js
@@ -15,17 +15,30 @@ function encrypt(input) {
 	return keyBase64;
 }
 
-function savePassword(pwd) {
+function password(pwd, event, done) {
 	const encryptedPassword = encrypt(pwd);
 	const userDataPath = app.getPath('userData');
 	const filePath = path.join(userDataPath, 'password.FLE');
 	
 	try {
-		fs.writeFileSync(filePath, encryptedPassword, 'utf-8');
-		console.log(`Password saved to ${filePath}`);
+		if (fs.existsSync(filePath)) {
+			const storedPassword = fs.readFileSync(filePath, 'utf-8');
+			if (storedPassword === encryptedPassword) {
+				console.log('Password matches the stored password.');
+				done(true);
+			} else {
+				console.log('Password does not match the stored password.');
+				done(false);
+			}
+		} else {
+			fs.writeFileSync(filePath, encryptedPassword, 'utf-8');
+			console.log(`Password saved to ${filePath}`);
+			done(true);
+		}
 	} catch (error) {
 		console.error(`Failed to save password: ${error.message}`);
+		done(false);
 	}
 }
 
-module.exports = { encrypt, savePassword };
+module.exports = { encrypt, password };
